+++ EchoAppInstaller.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30611.23
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UWPPackageInstaller", "UWPPackageInstaller\UWPPackageInstaller.csproj", "{9DCD1469-A8F6-4037-B50A-D37EBA91E7B8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EchoLensInstaller", "EchoLensInstaller\EchoLensInstaller.csproj", "{9DCD1469-A8F6-4037-B50A-D37EBA91E7B8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F1F0D75A-8A44-4DD2-9F4E-86938A8765D4}"
 	ProjectSection(SolutionItems) = preProject
