+++ EchoLensInstaller/App.xaml
@@ -1,8 +1,8 @@
 ﻿<!-- ReSharper disable once Xaml.RedundantNamespaceAlias -->
 <Application
-    x:Class="UWPPackageInstaller.App"
+    x:Class="EchoLensInstaller.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:UWPPackageInstaller">
+    xmlns:local="using:EchoLensInstaller">
 
 </Application>
