+++ EchoLensInstaller/App.xaml.cs
@@ -7,7 +7,7 @@ using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
 using JetBrains.Annotations;
 
-namespace UWPPackageInstaller
+namespace EchoLensInstaller
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
