+++ EchoLensInstaller/Assets/Square150x150Logo.scale-100.png
