+++ EchoLensInstaller/Assets/Square44x44Logo.altform-unplated_targetsize-256.png
