+++ EchoLensInstaller/Assets/Square44x44Logo.targetsize-256.png
