+++ EchoLensInstaller/Assets/Square44x44Logo.targetsize-32.png
