+++ EchoLensInstaller/Package.appxmanifest
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest" xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10" xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities" IgnorableNamespaces="uap mp rescap">
-  <Identity Name="EchoPackageInstaller" Publisher="CN=Echo Installer" Version="1.0.7.0" />
+  <Identity Name="EchoPackageInstaller" Publisher="CN=Echo Installer" Version="1.0.1.0" />
   <mp:PhoneIdentity PhoneProductId="c6f5e411-49d5-41cc-9658-3369f9e41ff5" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
   <Properties>
     <DisplayName>Echo Installer</DisplayName>
@@ -14,7 +14,7 @@
     <Resource Language="x-generate" />
   </Resources>
   <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="UWPPackageInstaller.App">
+    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="EchoLensInstaller.App">
       <uap:VisualElements DisplayName="Echo Installer" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Echo App Installer for HoloLens" BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png">
         </uap:DefaultTile>
