+++ EchoLensInstaller/Properties/AssemblyInfo.cs
@@ -4,11 +4,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("UWPPackageInstaller")]
+[assembly: AssemblyTitle("EchoLensInstaller")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Equinor ASA")]
-[assembly: AssemblyProduct("UWPPackageInstaller")]
+[assembly: AssemblyProduct("EchoLensInstaller")]
 [assembly: AssemblyCopyright("MIT License")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
